Overhaul testing arena, fix rendering artifacts, and resolve GPU race conditions
Features & Tweaks:
* GameplayScene: Expanded the testing arena to 200x200 and added angled
  ramps and procedural circular poles (using a newly generated cylinder mesh)
  to better test physics and rendering.
* GameplayScene: Added a red capsule-shaped health bar to the top left
  of the screen using ImGui's background draw list.
* main3D.hlsl: Toned down the Image-Based Lighting (IBL) ambient
  contribution by multiplying the skybox sample by 0.3f, preventing the
  scene from looking washed out.

Rendering & Physics Bug Fixes:
* DirectX12Renderer: Fixed a critical CPU-GPU race condition.
  `m_ObjectDataPool` and `m_MaterialDataPool` are now ring-buffered
  per-frame (`s_FrameCount`). Previously, the CPU would overwrite UBO
  data while the GPU was still executing the previous frame, causing
  severe visual artifacting (such as materials flashing or meshes
  stretching across the screen).
* GameplayScene: Fixed a frustum culling bug where AABB extents did
  not account for object rotation. Rotated meshes (like the new ramps)
  would incorrectly pop out of existence at screen edges. The absolute
  values of the model matrix are now properly applied to the AABB.
* main3D.hlsl: Drastically reduced the manual shadow depth bias from
  0.005 to 0.0005 to fix severe "Peter Panning" (shadows detaching
  from objects).
* DirectX12Pipeline: Changed the shadow pass `CullMode` from BACK to
  NONE to ensure unclosed block geometry casts proper shadows.

Code Maintenance:
* Resolved multiple C4244 and C4267 warnings (size_t/int/float conversions)
  across GameplayScene.cpp, Animation.hpp, and ResourceManager.cpp to
  ensure a clean build.

+++ include/Engine/Graphics/Animation.hpp
@@ -31,7 +31,7 @@ namespace VECTOR {
 
         Bone* FindBone(const std::string& name);
 
-        inline float GetTicksPerSecond() const { return m_TicksPerSecond; }
+        inline float GetTicksPerSecond() const { return static_cast<float>(m_TicksPerSecond); }
         inline float GetDuration() const { return m_Duration; }
         inline const AssimpNodeData& GetRootNode() const { return m_RootNode; }
         inline const std::map<std::string, BoneInfo>& GetBoneIDMap() const { return m_BoneInfoMap; }
