+++ src/Engine/Core/ResourceManager.cpp
@@ -33,7 +33,7 @@ namespace VECTOR {
         std::string key = fontPath + "_" + std::to_string(fontSize);
 
         if (m_Fonts.find(key) == m_Fonts.end()) {
-            TTF_Font* font = TTF_OpenFont(fontPath.c_str(), fontSize);
+            TTF_Font* font = TTF_OpenFont(fontPath.c_str(), static_cast<float>(fontSize));
             if (!font) {
                 VECTOR_LOG_WARN(std::string("Failed to load font: ") + fontPath + "! SDL_Error: " + SDL_GetError());
                 return nullptr;
